services/drug-publication-pipeline/pyproject.toml
[project]
name = "drug-publication-pipeline"
version = "0.1.0"
description = "Drug Publication Data Pipeline"
authors = [
    {name = "Yves OUATTARA",email = "yves.ouattara@live.fr"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "pandas (>=2.3.0,<3.0.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "flask (==2.2.5)",
    "flask-appbuilder (==4.5.3)",
    "flask-babel (==2.0.0)",
    "flask-jwt-extended (==4.7.1)",
    "flask-limiter (==3.11.0)",
    "flask-login (==0.6.3)",
    "flask-sqlalchemy (==2.5.1)",
    "flask-session (==0.5.0)",
    "flask-wtf (==1.2.2)",
    "jinja2 (==3.1.6)",
    "grpcio (==1.71.0)",
    "gunicorn (==23.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "pyarrow (>=20.0.0,<21.0.0)"
]

[tool.poetry]
packages = [{include = "drug_publication_pipeline", from = "src"}]

[tool.poetry.group.airflow.dependencies]
apache-airflow = {version = "3.0.1", python = ">=3.12,<3.13"}
apache-airflow-providers-common-sql = {version = "1.27.0", python = ">=3.12,<3.13"}
apache-airflow-providers-ftp = {version = "3.12.3", python = ">=3.12,<3.13"}
apache-airflow-providers-http = {version = "5.2.2", python = ">=3.12,<3.13"}
apache-airflow-providers-imap = {version = "3.8.3", python = ">=3.12,<3.13"}
apache-airflow-providers-slack = {version = "9.0.5", python = ">=3.12,<3.13"}
apache-airflow-providers-sqlite = {version = "4.0.2", python = ">=3.12,<3.13"}

[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
black = "^25.1.0"
flake8 = "^7.2.0"
pylint = "^3.3.7"
mypy = "^1.16.0"
pytest = "^8.4.0"
pandas-stubs = "^2.2.3.250527"
pyhamcrest = "^2.1.0"
pytest-mock = "^3.14.1"

[tool.pylint.format]
max-line-length = "88"

[tool.isort]
# following black's doc for compatibility
# See https://black.readthedocs.io/en/stable/guides/using_black_with_other_tools.html#isort
profile = "black"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
